+++ planner_lib/accounts/constants.py
@@ -1,0 +1,3 @@
+""" Constants for accounts"""
+class AccountPermissions:
+    ADMIN = 'admin'